began moving functions from main into statemachine
