+++ shepherd_bms/src/main.cpp
@@ -9,150 +9,14 @@
 #include <LTC68041.h>
 #include "segment.h"
 #include "compute.h"
-#include "calcs.h"
 #include "datastructs.h"
 #include "analyzer.h"
+#include "stateMachine.h"
 
 WDT_T4<WDT1> wdt;
 
 AccumulatorData_t *prevAccData = nullptr;
 
-uint32_t bmsFault = FAULTS_CLEAR;
-
-uint16_t overVoltCount = 0;
-uint16_t underVoltCount = 0;
-uint16_t overCurrCount = 0;
-uint16_t chargeOverVolt = 0;
-uint16_t overChgCurrCount = 0;
-uint16_t lowCellCount = 0;
-
-/**
- * @brief Algorithm behind determining which cells we want to balance
- * @note Directly interfaces with the segments
- * 
- * @param bms_data 
- */
-void balanceCells(AccumulatorData_t *bms_data)
-{
-	bool balanceConfig[NUM_CHIPS][NUM_CELLS_PER_CHIP];
-
-	// For all cells of all the chips, figure out if we need to balance by comparing the difference in voltages
-    for(uint8_t chip = 0; chip < NUM_CHIPS; chip++)
-    {
-		for(uint8_t cell = 0; cell < NUM_CELLS_PER_CHIP; cell++)
-		{
-			uint16_t delta = bms_data->chipData[chip].voltageReading[cell] - (uint16_t)bms_data-> minVoltage.val;
-			if(delta > MAX_DELTA_V * 10000)
-				balanceConfig[chip][cell] = true;
-			else
-				balanceConfig[chip][cell] = false;
-        }
-    }
-	#ifdef DEBUG_CHARGING
-	Serial.println("Cell Balancing:");
-	for(uint8_t c = 0; c < NUM_CHIPS; c++)
-    {
-        for(uint8_t cell = 0; cell < NUM_CELLS_PER_CHIP; cell++)
-        {
-			Serial.print(balanceConfig[c][cell]);
-			Serial.print("\t");
-		}
-		Serial.println();
-	}
-	#endif
-
-	segment.configureBalancing(balanceConfig);
-}
-
-/**
- * @brief Returns if we want to balance cells during a particular frame
- * 
- * @param bmsdata 
- * @return true 
- * @return false 
- */
-bool balancingCheck(AccumulatorData_t *bmsdata)
-{
-	if (!compute.isCharging()) return false;
-	if (bmsdata->maxTemp.val > MAX_CELL_TEMP_BAL) return false;
-	if (bmsdata->maxVoltage.val <= (BAL_MIN_V * 10000)) return false;
-	if(bmsdata->deltVoltage <= (MAX_DELTA_V * 10000)) return false;
-
-	return true;
-}
-
-/**
- * @brief Returns if we want to charge cells during a particular frame
- * 
- * @param bmsdata 
- * @return true 
- * @return false 
- */
-bool chargingCheck(AccumulatorData_t *bmsdata)
-{
-	static Timer chargeTimeout;
-
-	if(!chargeTimeout.isTimerExpired()) return false;
-	if(!compute.isCharging()) return false;
-	if(bmsdata->maxVoltage.val >= (MAX_CHARGE_VOLT * 10000))
-	{
-		chargeOverVolt++;
-		if (chargeOverVolt > 100) 
-		{
-			chargeTimeout.startTimer(CHARGE_TIMEOUT);
-			return false;
-		}
-	} 
-	else 
-	{
-		chargeOverVolt = 0;
-	}
-
-	return true;
-}
-
-void broadcastCurrentLimit(AccumulatorData_t *bmsdata)
-{
-	// States for Boosting State Machine
-	static enum
-	{
-		BOOST_STANDBY,
-		BOOSTING,
-		BOOST_RECHARGE
-	}BoostState;
-
-	static Timer boostTimer;
-	static Timer boostRechargeTimer;
-
-	//Transitioning out of boost
-	if(boostTimer.isTimerExpired() && BoostState == BOOSTING)
-	{
-		BoostState = BOOST_RECHARGE;
-		boostRechargeTimer.startTimer(BOOST_RECHARGE_TIME);
-	}
-	//Transition out of boost recharge
-	if(boostRechargeTimer.isTimerExpired() && BoostState == BOOST_RECHARGE)
-	{
-		BoostState = BOOST_STANDBY;
-	}
-	//Transition to boosting
-	if((bmsdata->packCurrent) > ((bmsdata->contDCL)*10) && BoostState == BOOST_STANDBY)
-	{
-		BoostState = BOOSTING;
-		boostTimer.startTimer(BOOST_TIME);
-	}
-
-	//Currently boosting
-	if(BoostState == BOOSTING || BoostState == BOOST_STANDBY)
-	{
-		compute.sendMCMsg(bmsdata->chargeLimit, min(bmsdata->dischargeLimit, bmsdata->contDCL * CONTDCL_MULTIPLIER));
-	}
-	//Currently recharging boost
-	else
-	{
-		compute.sendMCMsg(bmsdata->chargeLimit, min(bmsdata->contDCL, bmsdata->dischargeLimit));
-	}
-}
 
 #ifdef DEBUG_STATS
 
@@ -229,97 +93,6 @@ const void printBMSStats(AccumulatorData_t *accData)
 }
 
 #endif
-
-/**
- * @brief Returns any new faults or current faults that have come up
- * @note Should be bitwise OR'ed with the current fault status
- * 
- * @param accData 
- * @return uint32_t 
- */
-uint32_t faultCheck(AccumulatorData_t *accData)
-{
-	// FAULT CHECK
-	// Check for fuckies
-	uint32_t faultStatus = 0;
-
-	// Over current fault for discharge
-	if ((accData->packCurrent) > ((accData->dischargeLimit)*10)) 
-	{
-		overCurrCount++;
-		if (overCurrCount > 10) 
-		{ // 0.10 seconds @ 100Hz rate
-			faultStatus |= DISCHARGE_LIMIT_ENFORCEMENT_FAULT;
-		}
-	} else 
-	{
-		overCurrCount = 0;
-	}
-
-	// Over current fault for charge
-	if ((accData->packCurrent) < 0 && abs((accData->packCurrent)) > ((accData->chargeLimit)*10))
-	{
-		overChgCurrCount++;
-		if (overChgCurrCount > 100) 
-		{ // 1 seconds @ 100Hz rate
-			faultStatus |= CHARGE_LIMIT_ENFORCEMENT_FAULT;
-		}
-	} 
-	else 
-	{
-		overChgCurrCount = 0;
-	} 
-
-	// Low cell voltage fault
-	if (accData->minVoltage.val < MIN_VOLT * 10000) 
-	{
-
-		underVoltCount++;
-		if (underVoltCount > 900)
-		{ // 9 seconds @ 100Hz rate
-			faultStatus |= CELL_VOLTAGE_TOO_LOW;
-		}
-	} 
-	else 
-	{
-		underVoltCount = 0;
-	}
-
-	// High cell voltage fault
-	if (((accData->maxVoltage.val > MAX_VOLT * 10000) && digitalRead(CHARGE_DETECT) == HIGH) || (accData->maxVoltage.val > MAX_CHARGE_VOLT * 10000)) 
-	{ // Needs to be reimplemented with a flag for every cell in case multiple go over
-		overVoltCount++;
-		if (overVoltCount > 900) { // 9 seconds @ 100Hz rate
-			faultStatus |= CELL_VOLTAGE_TOO_HIGH;
-		}
-	} 
-	else 
-	{
-		overVoltCount = 0;
-	}
-
-	// High Temp Fault
-	if (accData->maxTemp.val > MAX_CELL_TEMP) {
-		faultStatus |= PACK_TOO_HOT;
-	}
-
-	// Extremely low cell voltage fault
-	if (accData->minVoltage.val < 900) 
-	{ // 90mV
-		lowCellCount++;
-		if (lowCellCount > 100) { // 1 seconds @ 100Hz rate
-			faultStatus |= LOW_CELL_VOLTAGE;
-		}
-	} 
-	else 
-	{
-		lowCellCount = 0;
-	}
-
-	return faultStatus;
-}
-
-
 
 void shepherdMain()
 {
